+++ packages/ui-vue-2/src/props.ts
@@ -166,6 +166,16 @@ export const selectProps = {
     default: defaults.loading
   },
 
+  showLoadingMessage: {
+    type: Boolean,
+    default: defaults.showLoadingMessage
+  },
+
+  showNoOptionsMessage: {
+    type: Boolean,
+    default: defaults.showNoOptionsMessage
+  },
+
   loadingMessage: {
     type: [String, Function],
     default: defaults.loadingMessage